% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{brightness_correlate}
\alias{brightness_correlate}
\title{Calculate brightness correlate}
\usage{
brightness_correlate(c, J, A_w, F_L)
}
\arguments{
\item{c}{Chromatic induction factor (non-negative)}

\item{J}{Lightness correlate \link{0, 100}}

\item{A_w}{Achromatic response for reference white}

\item{F_L}{Luminance adaptation factor (non-negative)}
}
\value{
Numeric brightness correlate value
}
\description{
Computes the brightness correlate (Q) which represents the absolute
perceived brightness of a color.
}
\keyword{internal}
