% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{calculate_response}
\alias{calculate_response}
\title{Calculate post-adaptation response}
\usage{
calculate_response(x, FL)
}
\arguments{
\item{x}{Input value}

\item{FL}{Luminance adaptation factor}
}
\value{
Adapted response value
}
\description{
Calculate post-adaptation response
}
\keyword{internal}
