% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{cam16ucs_distance}
\alias{cam16ucs_distance}
\title{Calculate CAM16-UCS color difference}
\usage{
cam16ucs_distance(ucs1, ucs2)
}
\arguments{
\item{ucs1, ucs2}{CAM16-UCS coordinates}
}
\value{
Numeric color difference value
}
\description{
Calculates the Euclidean distance in CAM16-UCS space between two colors
}
