% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{cam16ucs_distance_matrix}
\alias{cam16ucs_distance_matrix}
\title{Calculate all CAM16-UCS color differences for a set of colors}
\usage{
cam16ucs_distance_matrix(cam16ucs)
}
\arguments{
\item{cam16ucs}{Matrix of CAM16-UCS coordinates (n × 3)}
}
\value{
Matrix of pairwise color differences
}
\description{
Calculate all CAM16-UCS color differences for a set of colors
}
