% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{cam16ucs_matrix_to_hex}
\alias{cam16ucs_matrix_to_hex}
\title{Convert matrix of CAM16-UCS coordinates to hex colors}
\usage{
cam16ucs_matrix_to_hex(ucs, ...)
}
\arguments{
\item{ucs}{Matrix of CAM16-UCS coordinates (columns: J_p, a_M, b_M)}

\item{...}{Additional parameters passed to cam16ucs_to_hex()}
}
\value{
Character vector of hex color codes
}
\description{
Convert matrix of CAM16-UCS coordinates to hex colors
}
