% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciecam16.R
\name{CIECAM16}
\alias{CIECAM16}
\title{CIECAM16 Color Appearance Model Implementation}
\description{
A comprehensive implementation of the CIECAM16 color appearance model that provides
functions for color space conversions and color difference calculations. The package
supports:
\itemize{
\item Forward and inverse transformations between CIE XYZ and CIECAM16
\item Conversion between hex colors and CIECAM16
\item Color difference calculations in CAM16-UCS space
\item Various viewing condition parameters
}
}
\details{
The implementation follows the CIE TC 1-91 specification and includes support for:
\itemize{
\item Different viewing conditions (Average, Dim, Dark)
\item Chromatic adaptation
\item Complete set of color appearance correlates
}
}
\references{
Li C, Li Z, Wang Z, et al. (2017). Comprehensive color solutions: CAM16, CAT16,
and CAM16-UCS. Color Research & Application, 42(6), 703-718.

Gao C, Xiao K, Pointer M, Li C. (2024). The development of the CIECAM16 and
visualization of its domain and range. Color Research & Application, 49(1), 1-17.

Hellwig L, Fairchild MD. (2022). Revising CAM16-UCS. Color and Imaging Conference,
Proceedings of the Color and Imaging Conference, 2022(1), 277-280.

CIE 248:2022. The CIE 2016 Colour Appearance Model for Colour Management Systems:
CIECAM16. Vienna: CIE Central Bureau.

Morovic P, et al. (Package Maintainers). colour-science/colour: A Color Science
Package for Python. https://github.com/colour-science/colour
}
