% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{degree_of_adaptation}
\alias{degree_of_adaptation}
\title{Calculate degree of adaptation}
\usage{
degree_of_adaptation(F, L_A)
}
\arguments{
\item{F}{Numeric adaptation factor in \link{0, 1} range}

\item{L_A}{Numeric luminance of adapting field in cd/m²}
}
\value{
Numeric degree of adaptation in \link{0, 1} range
}
\description{
Computes the degree of adaptation for color appearance models
using the given adaptation factor and luminance.
}
\examples{
degree_of_adaptation(1.0, 318.31)
}
\keyword{internal}
