% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repulsion.R
\name{enforce_boundaries}
\alias{enforce_boundaries}
\title{Enforce LAB Space Boundaries}
\usage{
enforce_boundaries(points, boundaries)
}
\arguments{
\item{points}{A matrix of LAB coordinates where each row represents a color,
and columns represent the L, a, and b values respectively.}

\item{boundaries}{A list containing the LAB space boundaries, with elements
'l', 'a', and 'b', each a numeric vector of length 2
specifying the min and max values for the respective
channels.}
}
\value{
A matrix of LAB coordinates with the same dimensions as the input
'points' matrix, where each row represents a color, and the columns
represent the adjusted l, a, and b values that are within the
specified boundaries.
}
\description{
This function ensures that all colors remain within the valid LAB color
space boundaries. It adjusts the L, a, and b values of each color point
to ensure they fall within the specified minimum and maximum limits for each
channel.
}
\keyword{internal}
