% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_adjustment.R
\name{evaluate_cvd_change}
\alias{evaluate_cvd_change}
\title{Evaluate Changes in Color Vision Deficiency (CVD) Scores}
\usage{
evaluate_cvd_change(new_scores, old_scores)
}
\arguments{
\item{new_scores}{A numeric vector or list containing the new CVD scores}

\item{old_scores}{A numeric vector or list containing the original CVD scores
Must be the same length as new_scores}
}
\value{
A logical value indicating whether the changes represent an overall improvement
(TRUE) or not (FALSE)
}
\description{
This function compares two sets of CVD scores and determines if the changes represent
an improvement. It uses a weighted approach that gives more importance to improvements
in scores that were initially closer to the minimum score.
}
\details{
The function calculates relative improvements between old and new scores and weights
them based on their proximity to the minimum score in the original set. Scores closer
to the minimum receive higher weights in the final calculation.

The weighted improvement is considered positive if the weighted sum of relative
improvements is greater than 0.
}
\examples{
old_scores <- list(10, 15, 20)
new_scores <- list(8, 14, 19)
evaluate_cvd_change(new_scores, old_scores)

}
