% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{f_e_inverse}
\alias{f_e_inverse}
\title{Apply inverse non-linear response compression}
\usage{
f_e_inverse(x, F_L)
}
\arguments{
\item{x}{Numeric compressed value}

\item{F_L}{Luminance adaptation factor (non-negative)}
}
\value{
Numeric decompressed value
}
\description{
Applies the inverse non-linear response compression function used in
the CIECAM16 color appearance model.
}
\keyword{internal}
