% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{hex_cam16ucs_distance}
\alias{hex_cam16ucs_distance}
\title{Calculate CAM16-UCS color difference between hex colors}
\usage{
hex_cam16ucs_distance(hex1, hex2, ...)
}
\arguments{
\item{hex1, hex2}{Hex color codes}

\item{...}{Additional parameters passed to hex_to_cam16ucs()}
}
\value{
Numeric color difference value
}
\description{
Calculate CAM16-UCS color difference between hex colors
}
