% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{hex_cam16ucs_distance_matrix}
\alias{hex_cam16ucs_distance_matrix}
\title{Calculate pairwise CAM16-UCS color differences for hex colors}
\usage{
hex_cam16ucs_distance_matrix(colors, ...)
}
\arguments{
\item{colors}{Vector of hex color codes}

\item{...}{Additional parameters passed to hex_to_cam16ucs()}
}
\value{
Matrix of pairwise color differences
}
\description{
Calculate pairwise CAM16-UCS color differences for hex colors
}
