% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{hex_to_cam16ucs}
\alias{hex_to_cam16ucs}
\title{Convert hex colors to CAM16-UCS}
\usage{
hex_to_cam16ucs(hex, L_A = 100, Y_b = 20, surround = "average")
}
\arguments{
\item{hex}{Character vector of hex color codes}

\item{L_A}{Adapting luminance in cd/m²}

\item{Y_b}{Background relative luminance}

\item{surround}{Surround conditions}
}
\value{
Matrix of CAM16-UCS coordinates (J', a_M', b_M')
}
\description{
Convert hex colors to CAM16-UCS
}
