% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{hex_to_cam16ucs_matrix}
\alias{hex_to_cam16ucs_matrix}
\title{Convert hex colors to CAM16-UCS matrix}
\usage{
hex_to_cam16ucs_matrix(hex, ...)
}
\arguments{
\item{hex}{Character vector of hex color codes}

\item{...}{Additional parameters passed to hex_to_cam16ucs()}
}
\value{
Matrix of CAM16-UCS coordinates
}
\description{
Convert hex colors to CAM16-UCS matrix
}
