% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam16.R
\name{invert_response}
\alias{invert_response}
\title{Invert post-adaptation response}
\usage{
invert_response(x, FL)
}
\arguments{
\item{x}{Adapted value}

\item{FL}{Luminance adaptation factor}
}
\value{
Original value
}
\description{
Invert post-adaptation response
}
\keyword{internal}
