% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{lightness_correlate}
\alias{lightness_correlate}
\title{Calculate lightness correlate}
\usage{
lightness_correlate(A, A_w, c, z)
}
\arguments{
\item{A}{Achromatic response of the color}

\item{A_w}{Achromatic response for reference white}

\item{c}{Chromatic induction factor (non-negative)}

\item{z}{Scaling factor (non-negative)}
}
\value{
Numeric lightness correlate value \link{0, 100}
}
\description{
Computes the lightness correlate (J) which represents the perceived
brightness of a color relative to white.
}
\keyword{internal}
