% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{MATRIX_16}
\alias{MATRIX_16}
\title{Get CIECAM16 transformation matrix}
\usage{
MATRIX_16()
}
\value{
Numeric matrix (3x3) for XYZ to RGB transformation
}
\description{
Returns the 3x3 matrix used to transform from XYZ to RGB color space
in the CIECAM16 color appearance model.
}
\examples{
MATRIX_16()
}
\keyword{internal}
