% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{MATRIX_INVERSE_16}
\alias{MATRIX_INVERSE_16}
\title{Get inverse CIECAM16 transformation matrix}
\usage{
MATRIX_INVERSE_16()
}
\value{
Numeric matrix (3x3) for RGB to XYZ transformation
}
\description{
Returns the inverse of the 3x3 matrix used to transform from RGB to XYZ color space
in the CIECAM16 color appearance model.
}
\examples{
MATRIX_INVERSE_16()
}
\keyword{internal}
