% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{mean_bottom_quartile}
\alias{mean_bottom_quartile}
\title{Calculate Mean of Bottom Quartile}
\usage{
mean_bottom_quartile(distances)
}
\arguments{
\item{distances}{A numeric vector of distances.}
}
\value{
A numeric value representing the mean of the bottom quartile of the input distances.
}
\description{
This function calculates the mean of the bottom quartile of a given set of distances.
}
\details{
The function sorts the input distances, determines the size of the bottom quartile,
extracts the bottom quartile, and calculates its mean.
}
