% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{plot_state}
\alias{plot_state}
\title{Plot State Function}
\usage{
plot_state(points, base_colors = NULL, show_force_field = FALSE, ...)
}
\arguments{
\item{points}{A matrix or data frame containing the points to be plotted.
The first column should contain the point identifiers, the second column
should contain the x-coordinates, and the third column should contain the
y-coordinates.}

\item{base_colors}{A vector of base colors. If provided, the first \code{n_base}
points will be plotted with these colors.}

\item{show_force_field}{A logical value indicating whether to display the
force field visualization. Default is FALSE.}

\item{...}{Additional arguments passed to the \code{visualize_force_field} function.}
}
\description{
This function plots the state of a given set of points. It can optionally
display a force field visualization.
}
\details{
If \code{show_force_field} is TRUE, the function calculates and plots a force
field using the \code{visualize_force_field} function. The force field is displayed
using a color gradient and contour lines. If \code{show_force_field} is FALSE,
the function plots an empty plot with specified x and y limits.

The points are plotted on top of the force field or empty plot. If \code{base_colors}
is provided, the first \code{n_base} points are plotted with these colors and
larger point characters.
}
\examples{
\dontrun{
points <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), ncol = 3)
base_colors <- c("red", "blue")
plot_state(points, base_colors, show_force_field = TRUE)
}

}
