% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.huerd_evaluation}
\alias{print.huerd_evaluation}
\title{Print Method for huerd_evaluation Objects}
\usage{
\method{print}{huerd_evaluation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{huerd_evaluation}.}

\item{...}{Additional arguments (not used).}
}
\description{
Custom print method for evaluation results from \code{evaluate_palette}.
}
