% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.huerd_simulation_result}
\alias{print.huerd_simulation_result}
\title{Print Method for huerd_simulation_result Objects}
\usage{
\method{print}{huerd_simulation_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{huerd_simulation_result}.}

\item{...}{Additional arguments (not used).}
}
\description{
Custom print method for simulation results from \code{simulate_palette_cvd}.
}
