% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utility.R
\name{validate_color_input}
\alias{validate_color_input}
\title{Input validation for color values}
\usage{
validate_color_input(points, space = "lab", distance = FALSE)
}
\arguments{
\item{points}{Matrix of color values}

\item{space}{Character string indicating color space ("lab", "rgb", "xyz")}

\item{distance}{Logical indicating whether points are used for distance
calculation}
}
\value{
Validated matrix or error
}
\description{
Input validation for color values
}
\keyword{internal}
