% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_distances.R
\name{calculate_all_distances}
\alias{calculate_all_distances}
\title{Calculate color distances using CAM16-UCS}
\usage{
calculate_all_distances(
  colors,
  viewing_conditions = list(L_A = 100, Y_b = 20, surround = "average")
)
}
\arguments{
\item{colors}{Matrix of xyz color values (n × 3)}

\item{viewing_conditions}{List with viewing condition parameters}
}
\value{
List of distance vectors for original and CVD-simulated colors
}
\description{
Calculates color distances in CAM16-UCS space for both original and
CVD-simulated colors. Input colors should be in LAB space.
}
