% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{mean_bottom_quartile}
\alias{mean_bottom_quartile}
\title{Calculate Mean of Bottom Quartile}
\usage{
mean_bottom_quartile(distances)
}
\arguments{
\item{distances}{Numeric vector of distances}
}
\value{
Mean value of the bottom quartile
}
\description{
Helper function to calculate the mean of the lowest 25\% of values in a
vector.
}
